requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
click==8.1.3
    # via uvicorn
dramatiq[rabbitmq]==1.13.0
    # via -r requirements.in
fastapi==0.79.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via anyio
loguru==0.6.0
    # via -r requirements.in
pika==1.3.0
    # via dramatiq
prometheus-client==0.14.1
    # via dramatiq
pydantic==1.9.1
    # via fastapi
python-dotenv==0.20.0
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0
    # via uvicorn
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
uvicorn[standard]==0.18.2
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchfiles==0.16.1
    # via uvicorn
websockets==10.3
    # via uvicorn
